requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
certifi==2026.4.22
charset-normalizer==3.4.7
Deprecated==1.3.1
Django==6.0.4
django-anymail==15.0
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.17.1
et_xmlfile==2.0.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
openpyxl==3.1.5
packaging==26.2
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic_core==2.46.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
six==1.17.0
sqlparse==0.5.5
standardwebhooks==1.0.1
svix==1.92.2
types-Deprecated==1.3.1.20260408
types-python-dateutil==2.9.0.20260408
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
wrapt==2.1.2
